use while loop with delay for fetching from omnivore on loop
