+++ src/index.ts
@@ -1,209 +1,9 @@
 import '@logseq/libs'
-import {
-  BlockEntity,
-  LSPluginBaseInfo,
-  SettingSchemaDesc,
-} from '@logseq/libs/dist/LSPlugin'
 import * as icon from './images/icon.png'
-import { loadArticle, loadArticles } from './util'
-
-const settings: SettingSchemaDesc[] = [
-  {
-    key: 'api key',
-    type: 'string',
-    title: 'Enter Omnivore Api Key',
-    description: 'Enter Omnivore Api Key here',
-    default: '',
-  },
-  {
-    key: 'username',
-    type: 'string',
-    title: 'Enter Omnivore username',
-    description: 'Enter Omnivore username here',
-    default: '',
-  },
-  {
-    key: 'frequency',
-    type: 'number',
-    title: 'Enter sync with Omnivore frequency',
-    description:
-      'Enter sync with Omnivore frequency in minutes here or 0 to disable',
-    default: 60,
-  },
-]
-const delay = (t = 100) => new Promise((r) => setTimeout(r, t))
-let loading = false
-
-const fetchOmnivore = async (
-  apiKey: string,
-  username: string,
-  inBackground = false
-): Promise<void> => {
-  if (loading) return
-
-  if (!apiKey || !username) {
-    await logseq.UI.showMsg('Missing Omnivore username or api key', 'warning')
-
-    return
-  }
-
-  const pageName = 'Omnivore'
-  const blockTitle = '## 🔖 Articles'
-  const highlightTitle = '### 🔍 [[Highlights]]'
-  const fetchingTitle = '🚀 Fetching articles ...'
-
-  !inBackground && logseq.App.pushState('page', { name: pageName })
-
-  await delay(300)
-
-  loading = true
-  let targetBlock: BlockEntity | null = null
-  let lastFetchedAt = ''
-
-  try {
-    !inBackground && (await logseq.UI.showMsg('🚀 Fetching articles ...'))
-
-    let omnivorePage = await logseq.Editor.getPage(pageName)
-    if (!omnivorePage) {
-      omnivorePage = await logseq.Editor.createPage(pageName)
-    }
-    if (!omnivorePage) {
-      throw new Error('Failed to create page')
-    }
-
-    const pageBlocksTree = await logseq.Editor.getPageBlocksTree(pageName)
-    targetBlock = pageBlocksTree.length > 0 ? pageBlocksTree[0] : null
-    let lastUpdateAt = ''
-    if (targetBlock) {
-      const matches = targetBlock.content.match(
-        /(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):(\d{2}).(\d{3})Z/g
-      )
-      if (matches) {
-        lastUpdateAt = matches[0]
-      } else {
-        lastUpdateAt = logseq.settings?.['synced at'] as string
-      }
-      lastFetchedAt = lastUpdateAt
-      await logseq.Editor.updateBlock(targetBlock.uuid, fetchingTitle)
-    } else {
-      targetBlock = await logseq.Editor.appendBlockInPage(
-        pageName,
-        fetchingTitle
-      )
-    }
-    if (!targetBlock) {
-      throw new Error('block error')
-    }
-
-    const size = 100
-    let after = 0
-    /*eslint no-constant-condition: ["error", { "checkLoops": false }]*/
-    while (true) {
-      const [articles, hasNextPage] = await loadArticles(
-        apiKey,
-        after,
-        size,
-        lastUpdateAt
-      )
-
-      for (const { title, author, slug, description } of articles) {
-        const { labels, highlights, savedAt } = await loadArticle(
-          username,
-          slug,
-          apiKey
-        )
-
-        const content = `[${title}](https://omnivore.app/${username}/${slug})
-        collapsed:: true
-        author:: "${author}"
-        labels:: ${
-          labels
-            ? labels.map((l: { name: string }) => `[[${l.name}]]`).join(' ')
-            : 'null'
-        }
-        date:: ${new Date(savedAt).toDateString()}
-        > ${description}`
-
-        // remove existing block for the same article
-        const existingBlocks = await logseq.DB.q<BlockEntity>(`"${slug}"`)
-        if (existingBlocks && existingBlocks.length > 0) {
-          console.log(existingBlocks)
-          for (const block of existingBlocks) {
-            await logseq.Editor.removeBlock(block.uuid)
-          }
-        }
-
-        const articleBlock = await logseq.Editor.insertBlock(
-          targetBlock.uuid,
-          content,
-          { before: true, sibling: false }
-        )
-        if (!articleBlock) {
-          throw new Error('block error')
-        }
-
-        if (highlights?.length) {
-          const highlightBlock = await logseq.Editor.insertBlock(
-            articleBlock.uuid,
-            highlightTitle
-          )
-          if (!highlightBlock) {
-            throw new Error('block error')
-          }
-
-          for (const highlight of highlights) {
-            await logseq.Editor.insertBlock(
-              highlightBlock.uuid,
-              highlight.quote
-            )
-          }
-        }
-
-        // sleep for a second to avoid rate limit
-        await delay(1000)
-      }
-
-      if (!hasNextPage) {
-        break
-      }
-      after += size
-    }
-
-    !inBackground && (await logseq.UI.showMsg('🔖 Articles fetched'))
-  } catch (e) {
-    !inBackground &&
-      (await logseq.UI.showMsg('Failed to fetch articles', 'warning'))
-    console.error(e)
-  } finally {
-    loading = false
-    if (targetBlock) {
-      lastFetchedAt = new Date().toISOString()
-
-      await logseq.Editor.updateBlock(
-        targetBlock.uuid,
-        `${blockTitle} [:small.opacity-20 "fetched at ${lastFetchedAt}"]`
-      )
-
-      logseq.updateSettings({ 'synced at': lastFetchedAt })
-    }
-  }
-}
-
-const syncOmnivore = (
-  apiKey: string,
-  username: string,
-  frequency: number
-): number => {
-  let intervalID = 0
-  // sync every frequency minutes
-  if (frequency > 0) {
-    intervalID = setInterval(async () => {
-      await fetchOmnivore(apiKey, username, true)
-    }, frequency * 1000 * 60)
-  }
-
-  return intervalID
-}
+import { LSPluginBaseInfo } from '@logseq/libs/dist/LSPlugin'
+import { settingSchema } from './settingSchema'
+import { fetchArtifacts } from './fetchArtifacts'
+import { delay } from './util'
 
 /**
  * main entry
@@ -212,29 +12,22 @@ const syncOmnivore = (
 const main = async (baseInfo: LSPluginBaseInfo) => {
   console.log('logseq-omnivore loaded')
 
-  logseq.useSettingsSchema(settings)
+  logseq.useSettingsSchema(settingSchema)
 
   let apiKey = logseq.settings?.['api key'] as string
   let username = logseq.settings?.['username'] as string
   let frequency = logseq.settings?.frequency as number
-  let intervalID: number
+  let isLoading = false
 
   logseq.onSettingsChanged(() => {
     apiKey = logseq.settings?.['api key'] as string
     username = logseq.settings?.['username'] as string
-    const newFrequency = logseq.settings?.frequency as number
-    if (newFrequency !== frequency) {
-      if (intervalID) {
-        clearInterval(intervalID)
-      }
-      frequency = newFrequency
-      intervalID = syncOmnivore(apiKey, username, frequency)
-    }
+    frequency = logseq.settings?.frequency as number
   })
 
   logseq.provideModel({
     async loadOmnivore() {
-      await fetchOmnivore(apiKey, username)
+      await fetchArtifacts(apiKey, username)
     },
   })
 
@@ -254,11 +47,16 @@ const main = async (baseInfo: LSPluginBaseInfo) => {
     }
   `)
 
-  // fetch articles on startup
-  await fetchOmnivore(apiKey, username, true)
-
-  // sync every frequency minutes
-  intervalID = syncOmnivore(apiKey, username, frequency)
+  // Call fetchArtifacts on a loop with delay
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    if (!isLoading) {
+      isLoading = true
+      await fetchArtifacts(apiKey, username, true)
+      isLoading = false
+    }
+    await delay(frequency * 60 * 1000)
+  }
 }
 
 // bootstrap
